refactor: 패킷 사이즈, 명령 short로 변경

+++ server/ClientCommon/Packets.h
@@ -72,7 +72,7 @@ namespace ClientCommon
 		short level;
 		int   exp;
 		short x, y;
-		short placeId;
+		int placeId;
 	};
 
 	struct LoginFailResponse : BasePacket
